% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_norm_normalise_ext.r
\name{nm_apply_conditional_filter}
\alias{nm_apply_conditional_filter}
\title{Apply conditional filter to resample pool}
\usage{
nm_apply_conditional_filter(resample_df, conditional_on)
}
\description{
Apply conditional filter to resample pool
}
\keyword{internal}
