% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_io_gdas.r
\docType{data}
\name{nm_arl_gdas1_base_url}
\alias{nm_arl_gdas1_base_url}
\title{GDAS1 ARL archive base URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
nm_arl_gdas1_base_url
}
\description{
GDAS1 ARL archive base URL
}
\keyword{datasets}
