% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_backend_h2o.r
\name{nm_auto_target_mb}
\alias{nm_auto_target_mb}
\title{Determine Target Batch Size in MB}
\usage{
nm_auto_target_mb(model)
}
\arguments{
\item{model}{A trained H2O model object.}
}
\value{
Integer value in MB for batch sizing.
}
\description{
Determine Target Batch Size in MB
}
\keyword{internal}
