% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_utils_cache.r
\name{nm_cache_save}
\alias{nm_cache_save}
\title{Save a Result to Cache}
\usage{
nm_cache_save(cache_dir, key, result)
}
\arguments{
\item{cache_dir}{Character. Cache directory path.}

\item{key}{Character. Cache key (hex digest).}

\item{result}{R object to cache.}
}
\value{
Invisibly returns the file path.
}
\description{
Save a Result to Cache
}
\keyword{internal}
