% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_utils_prepare.r
\name{nm_check_data}
\alias{nm_check_data}
\title{Check Data Validity}
\usage{
nm_check_data(dt, feature_names, value)
}
\arguments{
\item{dt}{data.table to check}

\item{feature_names}{Character vector of feature columns}

\item{value}{Name of target column}
}
\value{
Cleaned and validated data.table
}
\description{
Validates the input data.table for modeling.
}
\keyword{internal}
