% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_utils_prepare.r
\name{nm_convert_ordered_to_factor}
\alias{nm_convert_ordered_to_factor}
\title{Convert Ordered Factors to Factors}
\usage{
nm_convert_ordered_to_factor(dt)
}
\description{
\code{nm_convert_ordered_to_factor} converts ordered factors to regular factors.
}
\keyword{internal}
