% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_io_eea.r
\docType{data}
\name{nm_eea_pollutant_codes}
\alias{nm_eea_pollutant_codes}
\title{EEA Pollutant Codes}
\format{
An object of class \code{integer} of length 6.
}
\usage{
nm_eea_pollutant_codes
}
\description{
Named integer vector mapping pollutant names to EEA/EIONET codes.
}
\keyword{datasets}
