% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_io_era5.r
\name{nm_era5_aq_variables_default}
\alias{nm_era5_aq_variables_default}
\title{Default Surface variables for Air-Quality Modelling (CDS names)}
\usage{
nm_era5_aq_variables_default
}
\description{
Default Surface variables for Air-Quality Modelling (CDS names)
}
