% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_utils_features.r
\name{nm_feature_importance_h2o}
\alias{nm_feature_importance_h2o}
\title{Extract Feature Importance Table from an H2O Model}
\usage{
nm_feature_importance_h2o(model, verbose = FALSE)
}
\arguments{
\item{model}{A trained H2O model object (class \code{H2OModel}).}

\item{verbose}{Logical flag to enable progress messages.}
}
\value{
A data frame containing variables and their importance metrics.
}
\description{
Helper function that interacts with H2O's backend to extract detailed variable importance.
}
