% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_backend_h2o.r
\name{nm_load_h2o}
\alias{nm_load_h2o}
\title{Load Saved H2O Model}
\usage{
nm_load_h2o(path = "./", filename = "automl", verbose = TRUE)
}
\arguments{
\item{path}{A string specifying the directory path where the model is saved. Default is './'.}

\item{filename}{A string specifying the name of the saved model file. Default is 'automl'.}

\item{verbose}{Should the function print log messages? Default is TRUE.}
}
\value{
The loaded H2O model object with its "backend" attribute correctly set.
}
\description{
\code{nm_load_h2o} loads a previously saved H2O model from disk
}
