% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_backend_lightgbm.r
\name{nm_load_lgb}
\alias{nm_load_lgb}
\title{Load Saved lightgbm Model}
\usage{
nm_load_lgb(path = "./", filename = "automl", verbose = TRUE)
}
\arguments{
\item{path}{A string specifying the directory path where the model is saved. Default is './'.}

\item{filename}{A string specifying the name of the saved model file. Default is 'automl'.}

\item{verbose}{Should the function print log messages? Default is TRUE.}
}
\value{
The loaded lightgbm Booster object with its "backend" attribute correctly set.
}
\description{
\code{nm_load_lgb} loads a previously saved lightgbm model from disk.
}
\keyword{internal}
