% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_utils_provenance.r
\name{nm_load_run}
\alias{nm_load_run}
\title{Load a Provenance Run from Disk}
\usage{
nm_load_run(path)
}
\arguments{
\item{path}{Character. File path to the .rds file, or the base path.}
}
\value{
An object of class \code{"nm_run"}.
}
\description{
Load an \code{nm_run} object previously saved with \code{nm_save_run}.
}
