% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_model_pdp.r
\name{nm_pdp_generic}
\alias{nm_pdp_generic}
\title{Compute PDP for Generic Models}
\usage{
nm_pdp_generic(
  df,
  model,
  var_list = NULL,
  training_only = TRUE,
  n_cores = NULL,
  grid_points = 50,
  quantile_range = c(0.01, 0.99),
  verbose = TRUE
)
}
\description{
Compute PDP for Generic Models
}
\keyword{internal}
