% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_utils_require.r
\name{nm_require}
\alias{nm_require}
\title{Require an R Package or Object with a Helpful Error Message}
\usage{
nm_require(module, hint = NULL)
}
\arguments{
\item{module}{A string specifying the package and, optionally, the object to load.
The R-idiomatic `::` is used as a separator.
Examples:
- `"dplyr"` (checks for the package)
- `"dplyr::select"` (retrieves the `select` function)}

\item{hint}{A string providing an installation hint (e.g., "install.packages('dplyr')")
to be included in the error message if the package is missing.}
}
\value{
If requesting a package, returns the package namespace. If requesting an
  object, returns the object itself.
}
\description{
\code{nm_require} checks for the availability of a package and stops with an
informative error if it's not installed. It can also be used to retrieve a
specific object from a package's namespace.
}
\examples{
\dontrun{
# Check for a package
nm_require("dplyr", hint = "install.packages('dplyr')")

# Get a specific function
select_fun <- nm_require("dplyr::select", hint = "install.packages('dplyr')")

# This will fail with a helpful error message if 'notapackage' is not installed
nm_require("notapackage", hint = "remotes::install_github('user/notapackage')")
}
}
