% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_backend_h2o.r
\name{nm_save_h2o}
\alias{nm_save_h2o}
\title{Save Trained H2O Model}
\usage{
nm_save_h2o(model, path = "./", filename = "automl", verbose = TRUE)
}
\arguments{
\item{model}{The trained H2O model object to save.}

\item{path}{A string specifying the directory path where the model will be saved.}

\item{filename}{A string specifying the desired filename for the saved model.}

\item{verbose}{Should the function print log messages? Default is `TRUE`.}
}
\value{
A string indicating the full path of the saved model.
}
\description{
`nm_save_h2o` saves a trained H2O model. It handles path normalization
and file renaming. This is the H2O-specific implementation.
}
\keyword{internal}
