% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_utils_provenance.r
\name{nm_save_run}
\alias{nm_save_run}
\title{Save a Provenance Run to Disk}
\usage{
nm_save_run(run, path, compress = TRUE)
}
\arguments{
\item{run}{An object of class \code{"nm_run"}.}

\item{path}{Character. File path (without extension).}

\item{compress}{Logical. Compress the RDS file.}
}
\value{
Invisibly, a named vector of saved file paths.
}
\description{
Persist an \code{nm_run} object to disk using \code{base::saveRDS}.
}
