% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_utils_metrics.r
\name{nm_Stats}
\alias{nm_Stats}
\title{Compute Statistics from DataFrame Columns}
\usage{
nm_Stats(df, mod = "value_predict", obs = "value", statistic = NULL)
}
\arguments{
\item{df}{Input data frame containing columns with predictions and observations.}

\item{mod}{Column name with model predictions. Default is "value_predict".}

\item{obs}{Column name with ground-truth observations. Default is "value".}

\item{statistic}{A character vector of metrics to compute. If NULL, a default set is used.}
}
\value{
A single-row data frame with the requested statistics.
}
\description{
\code{nm_Stats} computes requested statistics from two columns in a data frame.
}
