% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{my1}
\alias{my1}
\title{Manchester MY1 Monitoring Station Air Quality Dataset}
\format{
A data frame (or data.table) containing air quality and weather covariates.
}
\source{
Manchester Air Quality Network.
}
\usage{
my1
}
\description{
A sample dataset containing hourly concentrations of air pollutants and
meteorological variables from the MY1 monitoring station in Manchester.
}
\keyword{datasets}
