% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_utils_prepare.r
\name{nm_add_date_variables}
\alias{nm_add_date_variables}
\title{Add Date Variables}
\usage{
nm_add_date_variables(dt)
}
\description{
\code{nm_add_date_variables} adds date-related features using data.table.
}
\keyword{internal}
