% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_io_defra.r
\docType{data}
\name{nm_aurn_pollutant_codes}
\alias{nm_aurn_pollutant_codes}
\title{AURN Pollutant Codes}
\format{
An object of class \code{integer} of length 9.
}
\usage{
nm_aurn_pollutant_codes
}
\description{
Named integer vector mapping common pollutant names to EIONET codes.
}
\keyword{datasets}
