% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_pipeline_workflow.r
\name{nm_config_rolling}
\alias{nm_config_rolling}
\title{Create a Configuration for a Rolling Window Normalisation Run}
\usage{
nm_config_rolling(
  value,
  backend = "h2o",
  feature_names = NULL,
  variables_resample = NULL,
  split_method = "random",
  fraction = 0.75,
  model_config = NULL,
  n_samples = 300,
  window_days = 14,
  rolling_every = 7,
  seed = 7654321,
  n_cores = NULL,
  memory_save = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{value}{The target variable name as a string.}

\item{backend}{The modeling backend to use if a model needs to be trained (e.g., 'h2o').}

\item{feature_names}{The names of the features used for training and normalisation.}

\item{variables_resample}{The names of variables to be resampled during normalisation.}

\item{split_method}{Method for splitting data for model training (e.g., 'random').}

\item{fraction}{Proportion of data for training if a model is trained.}

\item{model_config}{A list of configuration parameters for model training.}

\item{n_samples}{Number of times to sample the data for normalisation.}

\item{window_days}{The size of the rolling window in days.}

\item{rolling_every}{The interval at which the rolling window is applied in days.}

\item{seed}{A random seed for reproducibility.}

\item{n_cores}{Number of CPU cores for parallel processing.}

\item{memory_save}{Logical flag for memory-efficient normalisation.}

\item{verbose}{Should the function print progress messages and logs?}
}
\value{
A named list of configuration options for `nm_rolling`.
}
\description{
This function creates a structured list of configuration options, specifically
for running a rolling window normalisation with `nm_rolling` or the `nm_run_workflow`
function with `mode = "rolling"`.
}
