% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_pipeline_workflow.r
\name{nm_config_single}
\alias{nm_config_single}
\title{Create a Configuration for a Single Normalisation Run}
\usage{
nm_config_single(
  value,
  backend = "h2o",
  feature_names = NULL,
  variables_resample = NULL,
  split_method = "random",
  fraction = 0.75,
  model_config = NULL,
  n_samples = 300,
  aggregate = TRUE,
  seed = 7654321,
  n_cores = NULL,
  memory_save = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{value}{The target variable name.}

\item{backend}{The modeling backend (e.g., 'h2o').}

\item{feature_names}{Character vector of feature names.}

\item{variables_resample}{Character vector of variables to resample.}

\item{split_method}{Method for splitting data.}

\item{fraction}{Proportion of data for training.}

\item{model_config}{List of model configuration parameters.}

\item{n_samples}{Number of samples for normalisation.}

\item{aggregate}{Logical flag to return aggregated results.}

\item{seed}{Random seed for reproducibility.}

\item{n_cores}{Number of CPU cores.}

\item{memory_save}{Logical flag for memory-efficient normalisation.}

\item{verbose}{Should the function print progress messages?}
}
\value{
A named list of configuration options for `nm_do_all`.
}
\description{
Create a Configuration for a Single Normalisation Run
}
