% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_pipeline_workflow.r
\name{nm_config_unc}
\alias{nm_config_unc}
\title{Create a Configuration for an Uncertainty Normalisation Run}
\usage{
nm_config_unc(..., n_models = 10, confidence_level = 0.95)
}
\arguments{
\item{...}{Parameters passed to `nm_config_single`.}

\item{n_models}{Number of models to train for uncertainty estimation.}

\item{confidence_level}{The confidence level for uncertainty estimation.}
}
\value{
A named list of configuration options for `nm_do_all_unc`.
}
\description{
Create a Configuration for an Uncertainty Normalisation Run
}
