% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_utils_features.r
\name{nm_extract_features}
\alias{nm_extract_features}
\title{Extract and Sort Feature Names by Importance from a Model}
\usage{
nm_extract_features(model, verbose = FALSE, ...)
}
\arguments{
\item{model}{The trained model object, which should have a 'backend' attribute.}

\item{verbose}{Should the function print log messages? Default is FALSE.}

\item{...}{Additional arguments to be passed to the backend-specific function,
such as `importance_ascending`.}
}
\value{
A vector of sorted feature names based on their importance.
}
\description{
\code{nm_extract_features} is a high-level wrapper that extracts feature names from a model.
It checks the model's 'backend' attribute to dispatch to the appropriate backend-specific function.
}
