% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_utils_features.r
\name{nm_extract_features_h2o}
\alias{nm_extract_features_h2o}
\title{Extract and Sort Feature Names by Importance from an H2O Model}
\usage{
nm_extract_features_h2o(model, importance_ascending = FALSE, verbose = FALSE)
}
\arguments{
\item{model}{The trained H2O model object.}

\item{importance_ascending}{A logical value for sorting order. Default is `FALSE`.}

\item{verbose}{Should the function print log messages? Default is FALSE.}
}
\value{
A vector of sorted feature names based on their importance.
}
\description{
\code{nm_extract_features_h2o} is the specific implementation for H2O models.
It uses H2O's `h2o.varimp` function to extract feature importance.
}
