% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_norm_normalise.r
\name{nm_generate_resampled}
\alias{nm_generate_resampled}
\title{Generate a Resampled Data Frame (data.table version)}
\usage{
nm_generate_resampled(df, resample_vars, replace, seed, resample_df)
}
\description{
Generate a Resampled Data Frame (data.table version)
}
\keyword{internal}
