% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_model_io.r
\name{nm_load_model}
\alias{nm_load_model}
\title{Load a saved model from a specified backend}
\usage{
nm_load_model(
  path = "./",
  filename = "automl",
  backend = "h2o",
  verbose = TRUE
)
}
\arguments{
\item{path}{A string specifying the directory path where the model is saved.}

\item{filename}{A string specifying the filename of the saved model.}

\item{backend}{A string specifying the model backend (e.g., 'h2o').}

\item{verbose}{Should the function print log messages? Default is TRUE.}
}
\value{
The loaded model object.
}
\description{
`nm_load_model` is a high-level wrapper that dispatches the model loading task.
}
