% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_norm_normalise.r
\name{nm_normalise}
\alias{nm_normalise}
\title{Normalise a Time Series Using a Trained Model}
\usage{
nm_normalise(df, model, ...)
}
\arguments{
\item{df}{The input data frame.}

\item{model}{The trained model object, which should have a 'backend' attribute.}

\item{...}{Additional arguments passed to the implementation function.}
}
\value{
A data frame containing the normalised results.
}
\description{
\code{nm_normalise} is a high-level wrapper that deweathers a time series. It
dispatches to a specific implementation based on the model's backend attribute.
}
