% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_norm_normalise.r
\name{nm_normalise_h2o}
\alias{nm_normalise_h2o}
\title{Normalise Data using an H2O Model (Implementation)}
\usage{
nm_normalise_h2o(
  df,
  model,
  feature_names,
  variables_resample = NULL,
  n_samples = 300,
  replace = TRUE,
  aggregate = TRUE,
  seed = 7654321,
  n_cores = NULL,
  weather_df = NULL,
  memory_save = FALSE,
  verbose = TRUE,
  n_retries = 3,
  model_features = NULL,
  cl = NULL
)
}
\arguments{
\item{df}{The input data frame or data.table.}

\item{model}{The trained model object, which should have a 'backend' attribute.}

\item{feature_names}{The names of the features used for training and decomposition.}

\item{variables_resample}{The names of variables to be resampled during normalisation.}

\item{n_samples}{Number of times to sample the data for normalisation.}

\item{replace}{Logical flag for resampling with or without replacement.}

\item{aggregate}{Logical flag to return aggregated results or individual samples.}

\item{seed}{A random seed for reproducibility.}

\item{n_cores}{Number of CPU cores to use for parallel processing.}

\item{weather_df}{A data frame to sample weather conditions from. If NULL, `df` is used.}

\item{memory_save}{Logical for memory-efficient normalisation.}

\item{verbose}{Should the function print progress messages and logs?}

\item{n_retries}{Number of times to retry a failed prediction batch.}

\item{model_features}{Optional character vector of features used by the model.}

\item{cl}{(Optional) An existing parallel cluster object. If provided, the function
will use it instead of creating and destroying its own.}
}
\value{
A data frame containing the normalised results.
}
\description{
`nm_normalise_h2o` is the specific implementation for normalising data using
a trained H2O model.
}
\keyword{(Implementation)}
\keyword{Data}
\keyword{H2O}
\keyword{Model}
\keyword{Normalise}
\keyword{an}
\keyword{internal}
\keyword{using}
