% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_model_pdp.r
\name{nm_pdp}
\alias{nm_pdp}
\title{Compute Partial Dependence Plots (PDP)}
\usage{
nm_pdp(df, model, var_list = NULL, verbose = TRUE, ...)
}
\arguments{
\item{df}{Data frame containing the input data.}

\item{model}{The trained model object, which should have a 'backend' attribute.}

\item{var_list}{A character vector of variables to compute PDP for. If NULL, PDPs
will be computed for all features used by the model.}

\item{verbose}{Logical. If TRUE, shows progress bars and logs.}

\item{...}{Additional arguments passed to the implementation function.}
}
\value{
A data frame with PDP results.
}
\description{
Compute Partial Dependence Plots (PDP)
}
