% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_model_pdp.r
\name{nm_pdp_h2o}
\alias{nm_pdp_h2o}
\title{Compute PDP for H2O Models}
\usage{
nm_pdp_h2o(df, model, var_list = NULL, training_only = TRUE, verbose = TRUE)
}
\description{
Compute PDP for H2O Models
}
\keyword{internal}
