% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_model_predict.r
\name{nm_predict}
\alias{nm_predict}
\title{Predict using a trained model}
\usage{
nm_predict(model, newdata, feature_names = NULL, verbose = FALSE, ...)
}
\arguments{
\item{model}{The trained model object, which should have a 'backend' attribute.}

\item{newdata}{A data frame containing the new data for prediction.}

\item{feature_names}{(Optional) A character vector of feature names to use.}

\item{verbose}{Should the function print log messages? Default is TRUE.}

\item{...}{Additional arguments to be passed to the backend-specific function.}
}
\value{
A numeric vector of predicted values.
}
\description{
\code{nm_predict} is a high-level wrapper that generates predictions from a trained model.
It checks the model's 'backend' attribute to dispatch to the appropriate backend-specific function.
}
