% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_utils_prepare.r
\name{nm_prepare_data}
\alias{nm_prepare_data}
\title{Prepare Data for Model Training}
\usage{
nm_prepare_data(
  df,
  value,
  feature_names,
  na_rm = TRUE,
  split_method = "random",
  fraction = 0.75,
  seed = 7654321,
  verbose = FALSE
)
}
\arguments{
\item{df}{The raw input data frame.}

\item{value}{A string indicating the target column in `df`.}

\item{feature_names}{A character vector of predictor columns to use.}

\item{na_rm}{If TRUE, rows with NA values are removed. If FALSE, they are imputed.}

\item{split_method}{A string for the data splitting strategy.}

\item{fraction}{A numeric value for the training fraction of the split.}

\item{seed}{An integer for the random seed.}

\item{verbose}{Should the function print log messages?}
}
\value{
A prepared data frame.
}
\description{
`nm_prepare_data` is a high-level wrapper that performs a series of data
preparation steps.
}
