% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_norm_rolling.r
\name{nm_rolling}
\alias{nm_rolling}
\title{Apply rolling window meteorological normalisation}
\usage{
nm_rolling(
  df = NULL,
  model = NULL,
  value = "value",
  feature_names = NULL,
  variables_resample = NULL,
  split_method = "random",
  fraction = 0.75,
  backend = "h2o",
  model_config = NULL,
  n_samples = 300,
  window_days = 14,
  rolling_every = 7,
  seed = 7654321,
  n_cores = NULL,
  memory_save = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{df}{Data frame containing the input data.}

\item{model}{Pre-trained model for normalisation. If not provided, a model will be trained.}

\item{value}{The target variable name as a string.}

\item{feature_names}{The names of the features used for normalisation.}

\item{variables_resample}{The names of variables to be resampled during normalisation.}

\item{split_method}{The method for splitting data into training and testing sets.}

\item{fraction}{The proportion of the data to be used for training.}

\item{backend}{The modeling backend to use if a model needs to be trained. Default is 'h2o'.}

\item{model_config}{A list containing configuration parameters for model training.}

\item{n_samples}{Number of times to sample the data for normalisation.}

\item{window_days}{The size of the rolling window in days.}

\item{rolling_every}{The interval at which the rolling window is applied in days.}

\item{seed}{A random seed for reproducibility.}

\item{n_cores}{Number of CPU cores to use for parallel processing.}

\item{memory_save}{Logical for memory-efficient normalisation.}

\item{verbose}{Should the function print progress messages and logs?}
}
\value{
A data frame containing the combined results from all rolling windows.
}
\description{
\code{nm_rolling} performs meteorological normalisation on data using a rolling window approach.
A single model is trained on the entire dataset, and then normalisation is applied to
sequential windows of that data.
}
