% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_model_io.r
\name{nm_save_model}
\alias{nm_save_model}
\title{Save a trained model}
\usage{
nm_save_model(model, path = "./", filename = "model", verbose = TRUE)
}
\arguments{
\item{model}{The trained model object to save. It must have a "backend" attribute.}

\item{path}{A string specifying the directory path where the model will be saved.}

\item{filename}{A string specifying the desired filename for the saved model.}

\item{verbose}{Should the function print log messages? Default is TRUE.}
}
\value{
A string indicating the path of the saved model.
}
\description{
\code{nm_save_model} is a high-level wrapper that saves a trained model.
It automatically detects the model type from its "backend" attribute and calls
the appropriate backend-specific saving function.
}
