% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_utils_prepare.r
\name{nm_split_into_sets}
\alias{nm_split_into_sets}
\title{Split Data into Training and Testing Sets}
\usage{
nm_split_into_sets(dt, split_method, fraction = 0.75, seed = 7654321)
}
\description{
\code{nm_split_into_sets} splits the data.table into training and testing sets.
}
\keyword{internal}
