% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_model_train.r
\name{nm_train_model}
\alias{nm_train_model}
\title{Train a model using a specified backend}
\usage{
nm_train_model(
  df,
  value = "value",
  variables = NULL,
  model_config = NULL,
  backend = "h2o",
  seed = 7654321,
  n_cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{df}{Input data frame containing the data to be used for training.}

\item{value}{The target variable name as a string. Default is "value".}

\item{variables}{Independent/explanatory variables used for training the model.}

\item{model_config}{A list containing configuration parameters for model training.}

\item{backend}{The modeling framework to use. Default is 'h2o'. Others can be added in the future.}

\item{seed}{A random seed for reproducibility. Default is 7654321.}

\item{n_cores}{Number of CPU cores to use for the model training. Default is system's total minus one.}

\item{verbose}{Should the function print progress messages? Default is TRUE.}
}
\value{
The trained model object from the specified backend.
}
\description{
\code{nm_train_model} is a high-level wrapper function that dispatches the training
task to a specific backend implementation based on the `backend` parameter.
The default backend is 'h2o'.
}
